backend/requirements.txt
# Q-Shield Backend Dependencies
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt,argon2]==1.7.4
cryptography==41.0.7
pyjwt==2.11.0

# Cryptography Analysis
pyOpenSSL==23.3.0
cryptography==41.0.7

# HTTP & Async
aiohttp==3.9.1
httpx==0.25.2
aiodns==3.1.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Logging & Monitoring
python-json-logger==2.0.7
python-dotenv==1.0.0

# PDF Generation
reportlab==4.0.9

# QR Codes
qrcode==7.4.2
pillow==10.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.12.0
isort==5.13.2
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0
click==8.1.7
